cool change
got off topic as was stuck on idea of creating an implementation. I added some logic to the server.js file and login page. Now the database has a role implementation on the mongoDB which is either staff, or client. Depending on what your role is it will route you to the correct page that is relevant for your profile